+++ src/constants/index.js
@@ -1,3 +1,5 @@
+const {ANDROID_ACTION_CHOICES, SCRIPT_PREFIX_ANDROID} = require("./android");
+const {IOS_ACTION_CHOICES, SCRIPT_PREFIX_IOS, IOS_DEVICE_PROPS, IOS_RUNTIME_PROPS} = require("./ios");
 const SCRIPT_PREFIX = "./src/shs/";
 
 const MAIN_ACTIONS = ["IOS", "ANDROID"];
@@ -9,71 +11,11 @@ const OS_TYPE_Q = {
   choices: MAIN_ACTIONS,
 };
 
-const IOS_RUNTIME_PROPS = {
-  type: "list",
-  name: "ios_runtime",
-  message: "Select an IOS Runtime",
-};
-
-const IOS_DEVICE_PROPS = {
-  type: "list",
-  name: "ios_device",
-  message: "Select an IOS Device",
-};
-
-const IOS_ACTION_CHOICES = [
-  {
-    key: "boot",
-    name: "Boot",
-    value: "boot",
-  },
-  {
-    key: "shutdown",
-    name: "Shutdown",
-    value: "shutdown",
-  },
-  {
-    key: "erase",
-    name: "Erase",
-    value: "erase",
-  },
-];
-
-const ANDROID_ACTION_CHOICES = [
-  {
-    key: "boot",
-    name: "Boot",
-    value: "boot",
-  },
-  {
-    key: "shutdown",
-    name: "Shutdown",
-    value: "shutdown",
-  },
-  {
-    key: "wipe",
-    name: "Wipe Data",
-    value: "wipe",
-  },
-  {
-    key: "debug",
-    name: "Open with debug flag",
-    value: "debug",
-  },
-  {
-    key: "logcat",
-    name: "Open with logcat flag",
-    value: "logcat",
-  },
-  {
-    key: "logcat_standalone",
-    name: "Standalone Logcat Window",
-    value: "standalone",
-  },
-];
 
 module.exports = {
   SCRIPT_PREFIX,
+  SCRIPT_PREFIX_ANDROID,
+  SCRIPT_PREFIX_IOS,
   MAIN_ACTIONS,
   OS_TYPE_Q,
   IOS_RUNTIME_PROPS,
