--- src/constants/iosRelated/constants.js
@@ -1,3 +1,0 @@
-export const IOS_KEYS = {};
-
-export const IOS_MESSAGES = {};
